// frontend/src/api/model/AdminSettingsDto.ts
/**
 * Generated by orval v8.7.0
 * Do not edit manually.
 * TeleHealth.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { AdminOperatingHourDto } from "./AdminOperatingHourDto";

export interface AdminSettingsDto {
  slug: string;
  clinicName: string;
  supportEmail: string;
  defaultAppointmentDurationMinutes: number;
  operatingHours: AdminOperatingHourDto[];
}