// frontend/src/api/model/AdminUpdateSettingsCommand.ts
/**
 * Generated by orval v8.7.0
 * Do not edit manually.
 * TeleHealth.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { AdminUpdateOperatingHourCommand } from "./AdminUpdateOperatingHourCommand";

export interface AdminUpdateSettingsCommand {
  clinicName: string;
  supportEmail: string;
  defaultAppointmentDurationMinutes: number;
  operatingHours: AdminUpdateOperatingHourCommand[];
}